requirements.txt
ebooklib
markdown
pillow
beautifulsoup4
